Backend-for-web/Backend-for-commercialweb/users/serializers.py
from rest_framework import serializers
from users.models import CustomUser
from rest_auth.registration.serializers import RegisterSerializer

class UserSerializer(serializers.ModelSerializer):
    class Meta:
        model = CustomUser
        fields = ('username', 'email', 'hoTen', 'soDt',)

class UpdateUserSerializer(serializers.ModelSerializer):
    class Meta:
        model = CustomUser
        fields = ('hoTen', 'soDt',)